requirements.txt
autograd==1.8.0
autograd-gamma==0.5.0
contourpy==1.3.0
cycler==0.12.1
ecos==2.0.14
feature_engine==1.9.3
fonttools==4.60.1
formulaic==1.2.1
importlib_resources==6.5.2
interface-meta==1.3.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
lifelines==0.30.0
MarkupSafe==3.0.3
matplotlib==3.9.4
narwhals==2.7.0
numexpr==2.10.2
numpy==2.0.2
osqp==1.0.4
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
pyarrow==21.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.5.2
scikit-survival==0.23.1
scipy==1.13.1
six==1.17.0
statsmodels==0.14.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
wrapt==1.17.3
zipp==3.23.0
